BUG#: 8624 TITLE: The cimserver build with slp coredumps on x86_64 DESCRIPTION: Changed the pointer typecasts from Uint32 to size_t

+++ src/slp/slp_client/src/cmd-utils/slp_client/y_filter.cpp
@@ -150,7 +150,7 @@
   int32 filterlex(void);
 
 int32 filterparse(void);
-void filter_close_lexer(uint32 handle);
+void filter_close_lexer(size_t handle);
 size_t filter_init_lexer(const char *s);
 
 /* have a place to put attributes and the filter while the parser is working */
@@ -1792,7 +1792,7 @@ lslpLDAPFilter *_lslpDecodeLDAPFilter(char *filter)
 {
 
   lslpLDAPFilter *temp = NULL;
-  uint32 lexer = 0;
+  size_t lexer = 0;
   PEGASUS_ASSERT(filter != NULL && strlen(filter));
   lslpInitFilterList();
   nesting_level = 1;
